Extended test cifs
Extended test cifs file choices

NOTE - there is an inconsistancy in the way that the reduced vectors
are chosen in the formulation of Slater et al. and what would be done
in a normal structure.
    - the difference is that normally you would pick one
crystallographic vector as your ‘x’, the slater hamiltonian assumes you
translated the diamond structure from a cubic structure of the same
reduced coordinates.